+++ EterTerminalDistribution/EterTerminalDistribution/CardType.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <cstdint>
 
-enum class CardType : uint8_t { MinionCard,ExplosionCard,HoleCard,MageCard,ElementalCard};
+enum class CardType : uint16_t { MinionCard,ExplosionCard,HoleCard,MageCard,ElementalCard};
